Comment widget, guestbook, Jay site, Slideshow
Huge update to the Rainbow parade website, the guestbook, logos and javbascript has all been added to the site. Started work on the Jay The Adventurer site and then completed the main framework for the picture slideshow. ver cool. probably need to refresh braion on wordpress, npm, jquery, and figma. nice.