Cleaning, added in Journal and Creations
Steamrolling through making many of the pages, starting with the creations, now over to the journal. Made assets when necessary for both pages, especially for the creations page. Once the framework for these sites are done, I can make the small pages that tie everything together. Cleaned up some miscellaneous code in the homepage, and updated the site map with links to the new pages.