Organized files and created temporary landing page
Organized all pages into it's own folder, there will be many more pages to come because of the gallery that needs to be made. Updated all pages to use the new javascript menus, and a failsafe just in case the user has javascript disabled.  Updated the sitemap to reflect new changes and site structure.